// frontend/src/app/api/model/createProjectInput.ts
/**
 * Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateProjectInput { 
    name?: string | null;
    description?: string | null;
    budget?: number | null;
    duration?: number | null;
    durationUnit?: number | null;
}